Reorganizacion de codigo luego del merge, modificacion menor de app.py Se cambian los campos: - nombre_completo -> nombre_cliente - dni -> cliente_id
